restructure unit tests. Fixes #51
