+++ tests/test_bytes_unicode.py
@@ -6,32 +6,23 @@
 # published by the Free Software Foundation
 #
 from __future__ import unicode_literals
-import sys
+
 import taglib
 from . import copy_test_file
 
 
-def test_bytes_filename(tmpdir):
-    """Ensure file can be opened if filename is provided as bytes object."""
-    f = copy_test_file('testöü.flac', tmpdir)
-    tf = taglib.File(f.encode('utf8'))
-    tf.close()
 
-
-def test_unicode_filename(tmpdir):
-    """Ensure file can be opened if filename is provided as bytes object."""
-    f = copy_test_file('testöü.flac', tmpdir)
-    if sys.version_info.major == 2:
-        f = unicode(f)
-    tf = taglib.File(f)
-    tf.close()
-
-
-def test_bytes_tags(tmpdir):
-    """Ensure bytes keys and values are accepted.
-
-    Update 2017-05-12: Use mixed-case tag values as a regression test for issue #33.
+def test_cyrillic_file_name(tmpdir):
+    """Motivated by https://github.com/supermihi/pytaglib/issues/28.
     """
+    copy_file = copy_test_file('Жбж.mp3', tmpdir)
+    tfile = taglib.File(copy_file)
+    tfile.tags['COMMENT'] = ['test']
+    tfile.save()
+    tfile.close()
+
+
+def test_accepts_bytes_keys_and_values(tmpdir):
     f = copy_test_file('rare_frames.mp3', tmpdir)
     tf = taglib.File(f)
     tf.tags[b'BYTES'] = [b'OnE', b'twO']
@@ -43,8 +34,7 @@ def test_bytes_tags(tmpdir):
     tf.close()
 
 
-def test_unicode_tags(tmpdir):
-    """Ensure unicode keys and values are accepted."""
+def test_accepts_unicode_keys_and_tags(tmpdir):
     f = copy_test_file('rare_frames.mp3', tmpdir)
     tf = taglib.File(f)
     tf.tags[u'UNICODE'] = [u'OnE', u'twO']
